move package to com.ktown4u.utils.
